requirements.txt
flask
requests
python-dotenv
streamlit
